requirements.txt
openai>=1.75.0
pillow>=10.4.0
streamlit>=1.39.0
python-dotenv>=1.0.1
